Moved slices to their own (now failing) tests
