+++ tests/test_cli.py
@@ -122,48 +122,3 @@ fandango:ERROR: Only found 0 perfect solutions, instead of the required 10
         self.assertEqual("", err)
         self.assertEqual("", out)
         self.assertEqual(0, code)
-
-    def test_startswith(self):
-        command = shlex.split(
-            """fandango fuzz -f tests/resources/twodigits.fan -n 1 -c '<start>.startswith("6")' --format=none --validate --random-seed 426912 --population-size 10"""
-        )
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_slice_1(self):
-        command = shlex.split(
-            """fandango fuzz -f tests/resources/twodigits.fan -n 1 -c '<start>[0] == "6"' --format=none --validate --random-seed 426912 --population-size 10"""
-        )
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_slice_2(self):
-        command = shlex.split(
-            """fandango fuzz -f tests/resources/twodigits.fan -n 1 -c '<start>[0:] == "66"' --format=none --validate --random-seed 426912 --population-size 10"""
-        )
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_slice_3(self):
-        command = shlex.split(
-            """fandango fuzz -f tests/resources/twodigits.fan -n 1 -c '<start>[1:] == "6"' --format=none --validate --random-seed 426912 --population-size 10"""
-        )
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
-
-    def test_slice_4(self):
-        command = shlex.split(
-            """fandango fuzz -f tests/resources/twodigits.fan -n 1 -c 'str(<start>[0:]) == "66"' --format=none --validate --random-seed 426912 --population-size 10"""
-        )
-        out, err, code = self.run_command(command)
-        self.assertEqual("", err)
-        self.assertEqual("", out)
-        self.assertEqual(0, code)
